Remove return false on 404 response
Hi Rhys, thanks for setting up the wrapper. It's currently serving me well to create listings on Etsy! 

I'm proposing to drop the return false on a 404 response. 
I was trying to create a listing with a wrong shipping_template_id, and instead of getting an error "404: Shipping template id is not found. I didn't got any error back regarding the Shipping template. 

Removing this got me the error message in Laravel I was looking for. 

Thanks again !